Adição de método para atualizar métodos longos após "alterar" a página de preferences/ValorMetodoLongoPreferencePage.java; Alteração nos métodos AdicionarAnaliseProjeto e RemoverAnaliseProjeto para chamda static; Alteração nos métodos para calcular os valores dos quartis (Q1, Q2 e Q3); Alteração de Tipo nos valores dos DadosComponentesArquiteturais (de double para int); Adição de lisra no ProviderModel para persistir os dados dos Componentes Arquiteturais (tabela de componentes arquiteturais); Alteração no AtualizadorInformacoesMetodoLongo para:    Inserir Valores medianos e limiares nos componenetes arquiteturais;    Filtar pelos valores obtidos no projeto de exemplo; Remoção de de logs (System.out.print); Criação de método na classe FiltrarMetodosLongos para filtrar os metodos longos por valores do projeto de exemplo (Faltando Implementar).

+++ src/com/cleverton/longmethoddetector/model/DadosComponentesArquiteturais.java
@@ -5,9 +5,9 @@ public class DadosComponentesArquiteturais {
 	private String extendsClass;
 	private String implementsArquitecture;
 	private String implementsAPIJava;
-	private Double mediana;
-	private Double primeiroQuartil;
-	private Double terceiroQuartil;
+	private int mediana;
+	private int primeiroQuartil;
+	private int terceiroQuartil;
 	
 	public String getExtendsClass() {
 		return extendsClass;
@@ -33,27 +33,27 @@ public class DadosComponentesArquiteturais {
 		this.implementsAPIJava = implementsAPIJava;
 	}
 	
-	public Double getMediana() {
+	public int getMediana() {
 		return mediana;
 	}
 	
-	public void setMediana(Double mediana) {
+	public void setMediana(int mediana) {
 		this.mediana = mediana;
 	}
 	
-	public Double getPrimeiroQuartil() {
+	public int getPrimeiroQuartil() {
 		return primeiroQuartil;
 	}
 	
-	public void setPrimeiroQuartil(Double primeiroQuartil) {
+	public void setPrimeiroQuartil(int primeiroQuartil) {
 		this.primeiroQuartil = primeiroQuartil;
 	}
 	
-	public Double getTerceiroQuartil() {
+	public int getTerceiroQuartil() {
 		return terceiroQuartil;
 	}
 	
-	public void setTerceiroQuartil(Double terceiroQuartil) {
+	public void setTerceiroQuartil(int terceiroQuartil) {
 		this.terceiroQuartil = terceiroQuartil;
 	}
